lawyer/requirements.txt
flask==2.3.3
werkzeug==2.3.7
flask-cors==3.0.10
google-generativeai==0.3.0